Add PuppetX::NetDev::EosApi class for API access
Without this patch a provider has no Ruby API to access the eAPI for
Arista.  This is a problem because each provider will need to gather
from the device and should not have to handle the HTTP REST API itself.
This patch addresses the problem by implementing a utility class.  Each
provider is expected to create an instance of this utility class and use
it to communicate with the Arista device.

The API also supports AF_UNIX domain sockets which is expected in a
future version of EOS.