Add EOS API methods for network_interface provider
Without this patch the EosApi class does not have instance methods to
support the network_interface provider for EOS.  This is a problem
because the provider needs to use the EosApi to configure network
interface attributes.  This patch addresses the problem by adding API
methods to configure the attributes required by the network_interface
provider.