Refactor EosApi using eapi_action for better errors
Without this patch each EosApi instance method implements its own error
checking and handling of the API response.  This is a problem because
there is a lot of code duplication that basically amounts to checking
for an 'error' key in the response, extracting the error message if so,
then raising an exception.

This patch addresses the problem by implementing a new method,
eapi_action, which takes a human readable string describing the high
level action being carried out in addition to the API commands.  The
method will check the API response for the 'error' key and raise a well
formatted, human readable error if there is an API error.

If there are no errors the value of the 'result' key is returned to the
caller.

This patch refactors all of the EosApi instance methods to use this
improved behavior in the eapi_action method.