Add format: text option to EosApi#eapi_action
Without this patch the EosApi cannot make an API call to obtain the
information returned by the `show interfaces switchport` command because
the EOS API does not implement this command when JSON data is being
returned in the caller.  This is a problem because Peter at Arista says
the `show interfaces switchport` command should be used to obtain all of
the information necessary to inspect the current state of virtual
interfaces (LAG's and SVI's).  This information is necessary to
implement the l2_interface provider for EOS.

This patch addresses the problem by adding an optional keyword argument
that makes an API call in text mode rather than JSON mode.  The API will
return the human readable output of the `show interfaces switchport`
command and this output will be parsed by an EosApi instance method.
When the EOS API supports this command in JSON mode it will be possible
to update the instance method to use structured JSON data without
changing the data returned to the provider implementation.