(maint) Add puppetmodule-netdev_stdlib gem dependency
Without this patch it is difficult to write provider spec tests because
the netdev_stdlib module is not located along the default module search
path when RSpec is initialize.  This results in
`Puppet::Type.type(:network_vlan)` returning `nil`.  This patch
addresses the problem by expressing the type library as a gem dependency
which is easily satisfied in an isolated and repeatable manner using
bundler.

With this patch, running the spec tests involves a greatly simplified
process of cloning the repository, using bundler to install all
dependencies (including puppet and netdev_stdlib), then executing rspec
spec.