(maint) Stub API methods for network_vlan provider
Without this patch the network_vlan specs are running against a live API
service running on a VM.  This is a problem because the specs are not
running in isolation and will not run in other environments.  This patch
addresses the problem by stubbing out the API calls for the network_vlan
provider.