(maint) Add EOS_{HOSTNAME,PORT,USERNAME,PASSWORD} API ENV variables
Without this patch the API REST endpoint is hard coded into the EosApi
class.  This is a problem because it is difficult to development and
test different behaviors when there is and is not an API present.  This
patch addresses the problem by checking the environment variables for
the API endpoint values.