(ARISTA-18) Fix network_interface failing on port channel interfaces
Without this patch the network_interface provider is incorrectly trying
to process Port-Channel interfaces as though they are ethernet
interfaces.  This is a problem because the Port-Channel interface does
not have some attributes, such as duplex, which was causing an error
while trying to list network interfaces when there are port channel
interfaces defined on the target switch.

This patch addresses the problem by processing only interfaces with a
hardware type of 'ethernet'.