(ARISTA-3) Manage port_channel speed, duplex, description
Without this patch the port_channel provider does not effectively
manage the speed, duplex and description attributes of the port channel
interface.  This patch addresses the problem by moving the
implementation methods for speed, duplex, and description from the
network_interface provider into the shared EosProviderMethods module and
re-using the methods for both the network_interface and the port_channel
providers.

The general approach to managing the speed and duplex settings are to do
so on each of the member interfaces.

When inspecting the current state of the system, the least common value
for speed and for duplex are used in an effort to ensure all member
interfaces remain consistent.  Given a channel group with 4 members,
if only one member has a different speed or duplex value then Puppet
will detect that and enforce the managed value across all members of
the channel group.