(ARISTA-3) Write port_channel minimum_links property
Without this patch the port_channel eos provider does not manage the
state of the minimum links property on the target device.  This patch
addresses the problem by implementing the set_portchannel_min_links
instance method on the EosApi class.  The patch updates the port_channel
provider to implement the minimum_links=(value) setter method to use
this API method.