(ARISTA-3) Read port_channel minimum_links property
Without this patch the port_channel eos provider does not inspect the
current state of the minimum links property.  This is a problem because
this property needs to be managed by Puppet.  This patch addresses the
problem by parsing the currently configured number of minimum links from
the running configuration of the switch.  If the text output from the
switch does not contain a min-links line for the port channel, we safely
assume there is a configured value of 0.