(maint) Move EOS_* configuration ENV vars to EosApi initializer
Without this patch the spec tests are not honoring the EOS_HOSTNAME,
EOS_USERNAME, EOS_PASSWORD and EOS_PORT environment variables.  This is
a problem because it makes it difficult to save fixture data from the
real API when working with the spec tests.

This patch addresses the problem by moving the logic that looks up
values from the environment out of the provider API helper method and
into the EosApi instance initializer.  This change ensure all
initialized instances of the API look up configuration values from the
environment in the same manner.