(ARISTA-3) Read flowcontrol port_channel values
Without this patch the EOS port_channel provider does not obtain the
current values for flowcontrol_send and flowcontrol_receive.  This is a
problem because we need to manage these values.  This patch addresses
the problem by iterating over the list of member interfaces and
using the least common value across the set of member interfaces.  This
approach is intended to identify minor discrepancies across the set of
member interfaces.