(maint) Use YAML instead of JSON for fixtures
Without this patch fixtures are not being saved correctly when the data
being saved is a Ruby Hash object with symbols for keys.  This is a
problem because the fixture save and restore process is not fully
isomorphic.  Consider the following:

    JSON.load(JSON.pretty_generate(foo: 1))

The actual result is `{"foo"=>1}` while the expected result is `{:foo =>
1}`.

This patch addresses the problem by saving YAML fixtures by default,
which correctly preserves Ruby native objects.  The patch modifies the
fixture() method to first check for the existence of a YAML fixture, and
load that, then fall back to a JSON fixture if YAML is not present.
This approach provides backwards compatibility while avoiding future
problems with saved fixtures.