(ARISTA-20) Add network_snmp provider spec tests for enable, contact, and location
Without this patch there are no spec tests for the provider itself.
This is a problem because there needs to be examples of how the provider
is expected to behave in order to support future refactoring efforts.
This patch addresses the problem by implementing behavior tests for each
of the provider methods that are currently implemented.