(NETDEV-18) Remove ensure property
Without this patch the ensure property is present on the provider.  This
is a problem because the type specification has a proposed change to
remove the ensure property because there will only ever be a single
resource of the network_snmp resource in a catalog.  This patch
addresses the problem by removing the ensure property from the provider
to match the proposed type specification.