(ARISTA-21) Inspect snmp_community resources
Without this patch there is no provider for the snmp_community type.
This patch partially addresses the problem by inspecting the current
state of all SNMP communities defined on the target device and
instantiating resources in puppet using the provider's `instances` class
method.

Remaining work is to augment the type to be ensurable and manage the
state of the target resources in addition to inspecting their current
state.