(maint) Fix snmp_community non-existent acl parsing bug
Without this patch a snmp_community resource with a non-existent ACL
defined is parsed as 'emanon (non-existent)' when it should be parsed as
'emanon'.  This is a problem because Puppet continually thinks the
resource is out of sync when it is in fact in sync.  This patch
addresses the problem by updating the regular expression that parses the
text output to deal with the presence of the trailing `(non-existent)`
information.