(ARISTA-21) Use snmp_community flush for management
Without this patch the snmp_community provider does not have the ability
to destroy resources on the target device, nor the ability to manage a
specific property of an existing resource.

This patch addresses the problem by implementing the provider flush
method.  The behavior of all the property setter methods is to set a key
for the property in the @property_flush instance variable.  The flush
method merges the @property_flush instance variable on top of the
@property_hash variable in order to preserve existing properties that
are not being managed by the applied catalog.