(ARISTA-23) Implement snmp_notification instances
Without this patch the snmp_notification type does not obtain all
existing snmp notification resources on the target device.  This is a
problem because we need all of the existing instances prior to managing
the resource specified in the Puppet configuration catalog.

This patch addresses the problem by implementing the provider instances
class method.  This method parses the output of the `show snmp trap`
command and returns a provider instance for each notification type.  For
example:

    snmp_notification { 'ospf nbr-state-change':
      enable => 'true',
    }
    snmp_notification { 'pim neighbor-loss':
      enable => 'false',
    }
    snmp_notification { 'snmp authentication':
      enable => 'true',
    }
    snmp_notification { 'snmp link-down':
      enable => 'true',
    }
    snmp_notification { 'snmp link-up':
      enable => 'true',
    }