(NETDEV-20) Fix composite namevar with resource purging
Without this patch the resources resource does not correctly purge
resources off the target system.  This is because the composite namevar
is implemented incorrectly.  This patch addresses the problem by
updating the composite namevar to use only 3 components, the name,
username (community), and port.  This change depends on a corresponding
change in the standard type library that eliminates the community
parameter.