(ARISTA-22) Add snmp_user provider
Without this patch there is no snmp_user provider for Arista EOS.  This
patch implements a working EOS snmp_user provider with a few caveats.
The caveats are that plaintext passwords are currently supported, but
the resource model does not work idempotently because Puppet cannot
inspect the plaintext password on the device, only a hashed value.
Furthermore, we are not sure how to specify a password hash value rather
than a plaintext password value using the eAPI.

In addition, the password will be used for both authentication and
privacy when configuring a V3 SNMP resource.

Based on advice from Peter at Arista this patch does not yet implement
the V3 private_key, localized_key, and enforce_privacy attributes.

The password and engine_id is retrieved from the target device by
listing the entire running configuration and parsing the text output.