(ARISTA-28) Update Gemfile and README for Eapi gem
Without this patch the provider module does not have the ability to use
the puppet_x_eos_eapi gem.  This is a problem because we need to remove
the included copy of the PuppetX::NetDev namespace and instead use the
API included in the Gem.

This patch addresses the problem by specifying the Gem dependency in the
Gemfile and including documentation about how to use the Gem from a Git
checkout in development situations.