(ARISTA-28) Remove PuppetX::NetDev::EosApi namespace
Without this patch the netdev_stdlib_eos module contains a vendored copy
of the PuppetX::NetDev::EosApi namespace.  This is a problem because
this namespace has been moved into a reusable Gem library named
puppet_x_eos_eapi.  This patch removes the included copy which results
in the Gem library being used.