(ARISTA-33) Create & destroy radius_server resources
Without this patch the radius_server provider does not have the ability
to create or destroy resources on the target device.  The following
error results from the attempt:

    $ puppet resource radius_server 10.11.12.15/1812/1813 ensure=present
    Error: Could not set 'present' on ensure: undefined method `create' for Radius_server[10.11.12.15/1812/1813]:Puppet::Type::Radius_server
    Error: Could not set 'present' on ensure: undefined method `create' for Radius_server[10.11.12.15/1812/1813]:Puppet::Type::Radius_server
    Wrapped exception:
    undefined method `create' for Radius_server[10.11.12.15/1812/1813]:Puppet::Type::Radius_server
    Error: /Radius_server[10.11.12.15/1812/1813]/ensure: change from absent to present failed: Could not set 'present' on ensure: undefined method `create' for Radius_server[10.11.12.15/1812/1813]:Puppet::Type::Radius_server
    radius_server { '10.11.12.15/1812/1813':
      ensure => 'absent',
    }

This patch addresses the problem by implementing the property setter
methods and deferring to the flush method to create the resource on the
target device.  This patch also implements the provider create and
destroy methods.