(ARISTA-30) Add tacacs_server_group provider
Without this patch there is no provider for the tacacs_server_group
type.  This patch addresses the problem by implementing the provider
fully.  Server groups are synchronized by first deleting the group and
then re-creating it from the complete list of managed servers for the
group.

Server groups look like:

    $ puppet resource tacacs_server_group
    tacacs_server_group { 'TAC-GR':
      ensure  => 'present',
      servers => ['1.2.3.4/4949', '1.2.3.5/49'],
    }
    tacacs_server_group { 'TAC-GR-EMPTY':
      ensure => 'present',
    }
    tacacs_server_group { 'TAC-GR-TRANS':
      ensure  => 'present',
      servers => ['1.2.3.5/50', '1.2.3.6/49', '1.2.3.6/50'],
    }
    tacacs_server_group { 'TAC-GR2':
      ensure  => 'present',
      servers => ['1.2.3.5/50', '1.2.3.6/49', '1.2.3.6/50'],
    }