(ARISTA-31) Add provider for tacacs_global type
Without this patch there is no provider for the tacacs_global type.
This patch implements the provider that configures the Tacacs global
default settings for the encryption key and timeout.  The provider does
not implement the retransmit_count property because this cannot be
configured at the global level on EOS.

This provide is largely a rip and replace of the radius_global provider.